refactor(context): move Eligibility Index context into CBV
one of the main motivations for the CBV refactor is to consolidate and
localize the data needed for views.
