+++ benefits/eligibility/context/__init__.py
@@ -1,4 +1,3 @@
-from .agency import eligibility_index
 from .flow import eligibility_start, eligibility_unverified
 
-__all__ = ["eligibility_index", "eligibility_start", "eligibility_unverified"]
+__all__ = ["eligibility_start", "eligibility_unverified"]
