+++ benefits/locale/en/LC_MESSAGES/django.po
@@ -6,7 +6,7 @@
 msgid ""
 msgstr ""
 "Report-Msgid-Bugs-To: https://github.com/cal-itp/benefits/issues \n"
-"POT-Creation-Date: 2025-07-30 18:48-0700\n"
+"POT-Creation-Date: 2025-08-06 15:07-0700\n"
 "Language: English\n"
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=UTF-8\n"
@@ -411,43 +411,6 @@ msgstr ""
 msgid "Which transit benefit would you like to enroll in?"
 msgstr ""
 
-#, python-format
-msgid ""
-"Cal-ITP doesn’t save any of your information. All CST transit benefits "
-"reduce fares by 50%% for bus service on fixed routes."
-msgstr ""
-
-#, python-format
-msgid ""
-"Cal-ITP doesn’t save any of your information. All MST transit benefits "
-"reduce fares by 50%% for bus service on fixed routes."
-msgstr ""
-
-#, python-format
-msgid ""
-"Cal-ITP doesn’t save any of your information. All Nevada County Connects "
-"transit benefits reduce fares by 50%% for bus service on fixed routes."
-msgstr ""
-
-#, python-format
-msgid ""
-"Cal-ITP doesn’t save any of your information. All SacRT transit benefits "
-"reduce fares by 50%% for bus service on fixed routes."
-msgstr ""
-
-#, python-format
-msgid ""
-"Cal-ITP doesn’t save any of your information. All SBMTD transit benefits "
-"reduce fares by 50%% for bus service on fixed routes."
-msgstr ""
-
-#, python-format
-msgid ""
-"Cal-ITP doesn’t save any of your information. All Ventura County "
-"Transportation Commission transit benefits reduce fares by 50%% for bus "
-"service on fixed routes."
-msgstr ""
-
 msgid "Get started with"
 msgstr ""
 
@@ -758,6 +721,43 @@ msgstr ""
 
 #, python-format
 msgid "Please reach out to %(short_name)s for assistance."
+msgstr ""
+
+#, python-format
+msgid ""
+"Cal-ITP doesn’t save any of your information. All CST transit benefits "
+"reduce fares by 50%% for bus service on fixed routes."
+msgstr ""
+
+#, python-format
+msgid ""
+"Cal-ITP doesn’t save any of your information. All MST transit benefits "
+"reduce fares by 50%% for bus service on fixed routes."
+msgstr ""
+
+#, python-format
+msgid ""
+"Cal-ITP doesn’t save any of your information. All Nevada County Connects "
+"transit benefits reduce fares by 50%% for bus service on fixed routes."
+msgstr ""
+
+#, python-format
+msgid ""
+"Cal-ITP doesn’t save any of your information. All SacRT transit benefits "
+"reduce fares by 50%% for bus service on fixed routes."
+msgstr ""
+
+#, python-format
+msgid ""
+"Cal-ITP doesn’t save any of your information. All SBMTD transit benefits "
+"reduce fares by 50%% for bus service on fixed routes."
+msgstr ""
+
+#, python-format
+msgid ""
+"Cal-ITP doesn’t save any of your information. All Ventura County "
+"Transportation Commission transit benefits reduce fares by 50%% for bus "
+"service on fixed routes."
 msgstr ""
 
 msgid "Your eligibility is confirmed! You’re almost there."
