+++ benefits/locale/es/LC_MESSAGES/django.po
@@ -6,7 +6,7 @@
 msgid ""
 msgstr ""
 "Report-Msgid-Bugs-To: https://github.com/cal-itp/benefits/issues \n"
-"POT-Creation-Date: 2025-07-30 18:48-0700\n"
+"POT-Creation-Date: 2025-08-06 15:07-0700\n"
 "Language: Español\n"
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=UTF-8\n"
@@ -564,57 +564,6 @@ msgstr "Ha cerrado sesión correctamente. ¡Gracias por usar Cal-ITP Benefits!"
 msgid "Which transit benefit would you like to enroll in?"
 msgstr "¿A qué beneficio de tránsito le gustaría inscribirse?"
 
-#, python-format
-msgid ""
-"Cal-ITP doesn’t save any of your information. All CST transit benefits "
-"reduce fares by 50%% for bus service on fixed routes."
-msgstr ""
-"Cal-ITP no almacena su información. Todos los beneficios de tránsito de CST "
-"reducen las tarifas de autobuses en un 50%%."
-
-#, python-format
-msgid ""
-"Cal-ITP doesn’t save any of your information. All MST transit benefits "
-"reduce fares by 50%% for bus service on fixed routes."
-msgstr ""
-"Cal-ITP no almacena su información. Todos los beneficios de tránsito de MST "
-"reducen las tarifas de autobuses en un 50%%."
-
-#, python-format
-msgid ""
-"Cal-ITP doesn’t save any of your information. All Nevada County Connects "
-"transit benefits reduce fares by 50%% for bus service on fixed routes."
-msgstr ""
-"Cal-ITP no almacena su información. Todos los beneficios de tránsito de "
-"Nevada County Connects reducen las tarifas de autobuses en un 50%%."
-
-#, python-format
-msgid ""
-"Cal-ITP doesn’t save any of your information. All SacRT transit benefits "
-"reduce fares by 50%% for bus service on fixed routes."
-msgstr ""
-"Cal-ITP no almacena su información. Todos los beneficios de tránsito de "
-"SacRT reducen las tarifas en un 50%% para el servicio de autobús en rutas "
-"fijas."
-
-#, python-format
-msgid ""
-"Cal-ITP doesn’t save any of your information. All SBMTD transit benefits "
-"reduce fares by 50%% for bus service on fixed routes."
-msgstr ""
-"Cal-ITP no almacena su información. Todos los beneficios de tránsito de "
-"SBMTD le ofrecen a los pasajeros tarifas de autobús reducidas."
-
-#, python-format
-msgid ""
-"Cal-ITP doesn’t save any of your information. All Ventura County "
-"Transportation Commission transit benefits reduce fares by 50%% for bus "
-"service on fixed routes."
-msgstr ""
-"Cal-ITP no almacena su información. Todos los beneficios de tránsito de "
-"Ventura County Transportation Commission reducen las tarifas de autobuses en "
-"un 50%%."
-
 msgid "Get started with"
 msgstr "Comience con"
 
@@ -974,6 +923,57 @@ msgstr "¡Está bien! Usted aún puede ser elegible para nuestro programa."
 #, python-format
 msgid "Please reach out to %(short_name)s for assistance."
 msgstr "Comuníquese con %(short_name)s para obtener ayuda."
+
+#, python-format
+msgid ""
+"Cal-ITP doesn’t save any of your information. All CST transit benefits "
+"reduce fares by 50%% for bus service on fixed routes."
+msgstr ""
+"Cal-ITP no almacena su información. Todos los beneficios de tránsito de CST "
+"reducen las tarifas de autobuses en un 50%%."
+
+#, python-format
+msgid ""
+"Cal-ITP doesn’t save any of your information. All MST transit benefits "
+"reduce fares by 50%% for bus service on fixed routes."
+msgstr ""
+"Cal-ITP no almacena su información. Todos los beneficios de tránsito de MST "
+"reducen las tarifas de autobuses en un 50%%."
+
+#, python-format
+msgid ""
+"Cal-ITP doesn’t save any of your information. All Nevada County Connects "
+"transit benefits reduce fares by 50%% for bus service on fixed routes."
+msgstr ""
+"Cal-ITP no almacena su información. Todos los beneficios de tránsito de "
+"Nevada County Connects reducen las tarifas de autobuses en un 50%%."
+
+#, python-format
+msgid ""
+"Cal-ITP doesn’t save any of your information. All SacRT transit benefits "
+"reduce fares by 50%% for bus service on fixed routes."
+msgstr ""
+"Cal-ITP no almacena su información. Todos los beneficios de tránsito de "
+"SacRT reducen las tarifas en un 50%% para el servicio de autobús en rutas "
+"fijas."
+
+#, python-format
+msgid ""
+"Cal-ITP doesn’t save any of your information. All SBMTD transit benefits "
+"reduce fares by 50%% for bus service on fixed routes."
+msgstr ""
+"Cal-ITP no almacena su información. Todos los beneficios de tránsito de "
+"SBMTD le ofrecen a los pasajeros tarifas de autobús reducidas."
+
+#, python-format
+msgid ""
+"Cal-ITP doesn’t save any of your information. All Ventura County "
+"Transportation Commission transit benefits reduce fares by 50%% for bus "
+"service on fixed routes."
+msgstr ""
+"Cal-ITP no almacena su información. Todos los beneficios de tránsito de "
+"Ventura County Transportation Commission reducen las tarifas de autobuses en "
+"un 50%%."
 
 msgid "Your eligibility is confirmed! You’re almost there."
 msgstr "¡Su elegibilidad está confirmada! Ya falta poco."
